Benchmarks for the OpenCRun runtime
The same benchmarks found in the 'src/opencl' folder are now located in
the 'src/opencrun' but, instead of the OCL_* environment variables, they
grab the preprocessor/linker flags from the OCRUN_* variables set by the
'configure' script when OpenCRun is enabled. In this way, both the
OpenCRun and the OpenCL generic runtime can be enabled and the OpenCRun
benchmark can be tailored for its platform without changing the
everything else.